Add files via upload
This script is used to track the PS5 console availability on amazon, as currently PS5's are scarce and rare for people to purchase due to a global shortage of semiconductors materials. The script uses the request library to constantly scrape the PS5 console page on amazon. Using the beautiful soup module as soon as the stock availability changed from "Currently unavailable." to "In stock." the automation bot created using the selenium library is triggered immediately and automates purchasing the PS5 by logging in, and clicking the buttons required to buy. Proxies were used as constant requests/web scraping from amazon with one constant ip address would result in a block by amazon. Hence, a proxy was used to rotate ip addresses constantly to give the illusion request were being made by different users across the country. 

- Libraries used:
- bs4
- requests
- datetime